+++ eventsdb/events.go
@@ -94,3 +94,24 @@ func (e SlashEvent) MarshalJSON() ([]byte, error) {
 		ValidatorPubKey: fmt.Sprintf("Mp%x", e.ValidatorPubKey),
 	})
 }
+
+type UnbondEvent struct {
+	Address         types.Address
+	Amount          []byte
+	Coin            types.CoinSymbol
+	ValidatorPubKey types.Pubkey
+}
+
+func (e UnbondEvent) MarshalJSON() ([]byte, error) {
+	return json.Marshal(struct {
+		Address         string `json:"address"`
+		Amount          string `json:"amount"`
+		Coin            string `json:"coin"`
+		ValidatorPubKey string `json:"validator_pub_key"`
+	}{
+		Address:         e.Address.String(),
+		Amount:          big.NewInt(0).SetBytes(e.Amount).String(),
+		Coin:            e.Coin.String(),
+		ValidatorPubKey: fmt.Sprintf("Mp%x", e.ValidatorPubKey),
+	})
+}